Make authorization token for users, WrongCredentialsException for throw errors. make password encoder.

+++ src/main/java/pl/luypawlowski/chessbackend/controller/UserController.java
@@ -5,10 +5,8 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import pl.luypawlowski.chessbackend.exception.UserExistsException;
-import pl.luypawlowski.chessbackend.model.NewEmailRequest;
-import pl.luypawlowski.chessbackend.model.NewLoginRequest;
-import pl.luypawlowski.chessbackend.model.NewPasswordRequest;
-import pl.luypawlowski.chessbackend.model.UserDto;
+import pl.luypawlowski.chessbackend.exception.WrongCredentialsException;
+import pl.luypawlowski.chessbackend.model.*;
 import pl.luypawlowski.chessbackend.service.UserService;
 
 import java.util.List;
@@ -22,9 +20,9 @@ public class UserController {
 
     @PostMapping
     public ResponseEntity saveUser(@RequestBody UserDto userDto) {
-        try{
+        try {
             return ResponseEntity.ok(userService.saveUser(userDto));
-        } catch (UserExistsException e){
+        } catch (UserExistsException e) {
             e.printStackTrace();
             return ResponseEntity.status(HttpStatus.CONFLICT).body(e.getMessage());
         }
@@ -44,23 +42,37 @@ public class UserController {
         return userService.getAllUsers();
     }
 
+    @PostMapping("/logIn")
+    public ResponseEntity logIn(@RequestBody UserLogInRequest userLogInRequest) {
+        try {
+            return ResponseEntity.ok(userService.logIn(userLogInRequest));
+        } catch (WrongCredentialsException e) {
+            e.printStackTrace();
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(e.getMessage());
+        }
+    }
+
     @DeleteMapping
-    public void deleteUserByLogin(@RequestParam("login") String login) {
-        userService.deleteUserByLogin(login);
+    public void deleteUserByLogin(@RequestParam("login") String login, @RequestHeader("Authorization") String authorization) {
+        if (userService.findUserAuthorizationToken(authorization,login)){
+            userService.deleteUserByLogin(login);
+        } else {
+            throw new WrongCredentialsException("Wrong data!");
+        }
     }
 
     @PutMapping("/{id}/new-password")
-    public UserDto editUserPassword(@RequestBody NewPasswordRequest newPasswordRequest, @PathVariable Long id){
+    public UserDto editUserPassword(@RequestBody NewPasswordRequest newPasswordRequest, @PathVariable Long id) {
         return userService.updateUserPassword(newPasswordRequest.getPassword(), id);
     }
 
     @PutMapping("/{id}/new-login")
-    public UserDto editUserLogin(@RequestBody NewLoginRequest newLoginRequest, @PathVariable Long id){
+    public UserDto editUserLogin(@RequestBody NewLoginRequest newLoginRequest, @PathVariable Long id) {
         return userService.updateUserLogin(newLoginRequest.getLogin(), id);
     }
 
     @PutMapping("/{id}/new-email")
-    public UserDto editUserEmail(@RequestBody NewEmailRequest newEmailRequest, @PathVariable Long id){
+    public UserDto editUserEmail(@RequestBody NewEmailRequest newEmailRequest, @PathVariable Long id) {
         return userService.updateUserEmail(newEmailRequest.getEmail(), id);
     }
 }
